+++ tests/testthat/helper-style.R
@@ -7,3 +7,8 @@ style_attr <- function(nodes, attr) {
   )
 }
 
+dash_array <- function(...) {
+  x <- xmlSVG(mini_plot(1:3, ..., type = "l"))
+  dash <- style_attr(xml_find_first(x, "//polyline"), "stroke-dasharray")
+  as.numeric(strsplit(dash, ",")[[1]])
+}