+++ NEWS.md
@@ -10,6 +10,9 @@
 
 * Greatly improves the performance of `svgstring()` (#58).
 
+* Clip paths now get a unique identifier to avoid collisions when
+  multiple plots are included in a document (#67).
+
 # svglite 1.1.0
 
 * Text metrics now converted from points to pixels (#45, #48) - this