+++ NEWS.md
@@ -13,6 +13,17 @@
 * Clip paths now get a unique identifier to avoid collisions when
   multiple plots are included in a document (#67).
 
+* Add `stringSVG()` for testing.
+
+* Text is now correctly scaled (#72, #59). See also the new vignette
+  on scaling.
+
+* The device functions gain `system_fonts` and `user_fonts`
+  arguments. See the new vignette on fonts for more information.
+
+* Raster graphics are now correctly cropped (#64) and handle
+  transparency properly.
+
 # svglite 1.1.0
 
 * Text metrics now converted from points to pixels (#45, #48) - this